# Faster prompt processing for IQ2_KS, IQ2_K, IQ2_K_R4